Update api validation logic

+++ src/main/java/com/istiak/timezone/controller/UserController.java
@@ -4,9 +4,11 @@ import com.istiak.timezone.config.JwtUtil;
 import com.istiak.timezone.constants.RestApiConstants;
 import com.istiak.timezone.model.AuthorityConstants;
 import com.istiak.timezone.model.UserDTO;
+import com.istiak.timezone.model.UserData;
 import com.istiak.timezone.model.UserResponse;
 import com.istiak.timezone.service.UserService;
 import com.istiak.timezone.service.UserValidationService;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -21,6 +23,7 @@ import javax.annotation.security.RolesAllowed;
 import java.util.Collections;
 import java.util.Map;
 
+@CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
 @RequestMapping("/api")
 public class UserController {
@@ -37,7 +40,7 @@ public class UserController {
     @Autowired
     private UserService  userService;
 
-    @RequestMapping(value="token",
+    @RequestMapping(value="signin",
             method = RequestMethod.POST,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Map<String,String>> authenticateUser(@RequestBody UserDTO userDTO) {
@@ -72,20 +75,28 @@ public class UserController {
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @RolesAllowed({AuthorityConstants.ADMIN})
-    public ResponseEntity<UserDTO> getSingleUser(@PathVariable Long id){
-        UserDTO userDTO = userService.getSingleUser(id);
-        if(userDTO == null) {
+    public ResponseEntity<UserData> getSingleUser(@PathVariable Long id){
+        UserData userData = userService.getSingleUser(id);
+        if(userData == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return ResponseEntity.ok(userDTO);
+        return ResponseEntity.ok(userData);
     }
 
-    @RequestMapping(value="users",
+    @RequestMapping(value={"users","signup"},
             method = RequestMethod.POST,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Void> createUser(@RequestBody UserDTO userDTO){
-        if(userValidationService.checkIfUserExists(userDTO)) {
-            return new ResponseEntity<>(HttpStatus.CONFLICT);
+    public ResponseEntity<String> createUser(@RequestBody UserDTO userDTO){
+        String errors = userValidationService.checkIfInvalidData(userDTO);
+        if(errors.length()>0) {
+            JSONObject errorMsg = new JSONObject();
+            errorMsg.put("msg", errors);
+            return new ResponseEntity<>(errorMsg.toString(),HttpStatus.BAD_REQUEST);
+        }
+        if(userValidationService.checkIfUserExists(userDTO.getUsername())) {
+            JSONObject errorMsg = new JSONObject();
+            errorMsg.put("msg", "User name already exists!");
+            return new ResponseEntity<>(errorMsg.toString(),HttpStatus.CONFLICT);
         }
         userService.createUser(userDTO);
         return new ResponseEntity<>(HttpStatus.CREATED);
@@ -95,10 +106,21 @@ public class UserController {
             method = RequestMethod.PUT,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @RolesAllowed({AuthorityConstants.ADMIN})
-    public ResponseEntity<Void> updateUser(@RequestBody UserDTO userDTO){
-        HttpStatus httpStatus = userValidationService.checkIfUserExists(userDTO)? null : HttpStatus.NOT_FOUND;
+    public ResponseEntity<?> updateUser(@RequestBody UserData userData){
+        if(userData == null || userData.getUsername()== null || userData.getUsername().length()==0) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        final String authenticatedUser = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        if(userData.getUsername().equalsIgnoreCase(authenticatedUser)) {
+            JSONObject errorMsg = new JSONObject();
+            errorMsg.put("msg", "Can't update own data!");
+            return new ResponseEntity<>(errorMsg.toString(),HttpStatus.BAD_REQUEST);
+        }
+        
+        HttpStatus httpStatus = userValidationService.checkIfUserExists(userData.getUsername())? null : HttpStatus.NOT_FOUND;
         if(httpStatus == null) {
-            httpStatus = userService.updateUser(userDTO) ? HttpStatus.OK : HttpStatus.BAD_REQUEST;
+            httpStatus = userService.updateUser(userData) ? HttpStatus.OK : HttpStatus.BAD_REQUEST;
         }
 
         return new ResponseEntity<>(httpStatus);
@@ -108,7 +130,16 @@ public class UserController {
             method = RequestMethod.DELETE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @RolesAllowed({AuthorityConstants.ADMIN})
-    public ResponseEntity<Void> deleteUser(@PathVariable String username){
+    public ResponseEntity<?> deleteUser(@PathVariable String username){
+        if(username == null || username.length()==0) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        final String authenticatedUser = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        if(username.equalsIgnoreCase(authenticatedUser)) {
+            JSONObject errorMsg = new JSONObject();
+            errorMsg.put("msg", "Can't delete own data!");
+            return new ResponseEntity<>(errorMsg.toString(),HttpStatus.BAD_REQUEST);
+        }
         HttpStatus httpStatus = userService.deleteUser(username)? HttpStatus.OK : HttpStatus.BAD_REQUEST;
         return new ResponseEntity<>(httpStatus);
     }
